app/requirements.txt
# FastAPI and Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.25
asyncpg==0.31.0
aiosqlite==0.18.0
psycopg2-binary==2.9.11
alembic==1.13.1

# Redis and Caching
redis==5.0.1
hiredis==2.3.2

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# API Communication Formats
# GraphQL support
strawberry-graphql[fastapi]==0.219.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.0.0

# Code Quality and Linting
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pylint==3.0.3

# Image Processing and File Handling
Pillow==10.1.0
python-magic==0.4.27
# For Windows, uncomment the following line
#python-magic-bin==0.4.14; sys_platform == "win32"

# Object Storage (MinIO / S3)
minio==7.2.0
boto3==1.34.3

# Malware Scanning
pyclamd==0.4.0

# Utilities
python-slugify[unidecode]==8.0.1
email-validator==2.1.0
phonenumbers==8.13.27

# Monitoring and Logging (Optional)
sentry-sdk[fastapi]==1.39.2